refactor(upload): extract view components and split into small methods
