+++ src/admin/upload/components/AttachmentDetailsView/style.scss
